Create game method
This method will be used to play a mastermind game. For the moment it
only uses a loop to iterate the rounds where the player has to guess
the code of the other one and display the turn. Now the class has a
variable to set the number of turns for a game.