Add start of game
In the game method, the computer creates a code that the human player
will guess.

To do that, first the Mastermind class now receives a computer and a
human player and methods to read and write their class variables.

Finally, the module now has a method to create a CODE class constant and
a method to read that variable.