Enter a code guess
Created the enter_code method to the Player class. This method asks the
player to enter a code of 6 colors; it creates a new class variable
called gues_code, it uses a while loop that ends when there are 6 valid
colors entered.

In the loop, it asks a color, if this color is valid and is not repeated
it is added to the player's guess. When it exits the loop, the method
returns this array.

Finally, in the game method, moved the turn to the top of the loop and
entered this new method and print the result when the player makes his
guess.