Create veryfiy_code method
This new method of the Mastermind class returns true when the player's
guess is equal to the other code and false otherwise.

In the game method, it has been added an if statement, that when it's
true, it informs that the player who made the guess wins the game and
exits the 12 turns loop.