Finish game method
First, there is now a new calculate_pegs method on the Mastermind class.

This method creates two new variables to show the pegs, then it loops
through the code to be guessed, compares each element and when a color
is guessed, it adds a red peg, and when it's the same color but a
different position it adds a white peg.

Then, in the game method, it is called the calculate_pegs method when
the player did not guessed the code and after the method execution it
tells the player how many red and white pegs it has.

It checks if the player did not guess the code after 12 turns and if
that happens it tells the player that the game is over and shows the
other player's code.