Create select_role method
The select_role method is used to let the human player choose the role
before the start of the game. Then it assigns the Human and Computer
class variables to new class variables for their roles in the game.

Also, in the game method it's replaced the calls to these methods using
these new class variables.