Create generate_combinations method
The generate_combinations method is a new method for the Computer class
that generates an array with all the possible combinations of colors;
this is used to try to guess the secret code in a game.

And finally, added a new role class variable to the Human class; this is useful
to define certain actions that only a computer will do in a game or
the actions for a human player.