Make computer's custom solving strategy First, the generate_code method on the Generable module now accepts a parameter with the number of colors that will generate.
The Computer class now has two new variables: code_guessed, that is
initialized with 6 nil values, it's used when the computer has guesed
the correct colors and possible_colors, this is a copy of COLORS from
the Generable module, used when the computer has guessed an incorrect
color, it eliminates it from this list.

On this same class, the generate_colors is overrided to use the new
possible_colors variable; also the enter_code has changed: it has a
parameter with the code the computer will guess each turn, this method
loops this array and when it founds a nil value, it genereates a color.
It ensures that it's not a duplicated color since the code is composed
of unique colors. When it assigns colors it assigns them to guess_code.

In the Mastermind class, the pegs are now initialized to 0 when an
instance is created, to avoid errors on the program. Also, the game
method now has a variable to store the guest's current color. In this
same method, the computer now enters a code each turn and when the game
tells the computer that it has 6 pegs in total it shuffles it's code.

Finally, on the calculate_pegs method it just helps the computer to
mantain colors when it justs obtains a red or white peg and deletes the
wrong colors from the possible_colors array.