Modify create_code method
Added a code variable with an empty list as a default value. Changed
the line that adds each color to the code instance variable.Modified the
condition to verify the code variable length instead of the code
instance variable. Added the code variable to return when the new
condition returns true.