Added game method
This method uses an infinite loop to iterate through each turn of a
mastermind game. It uses the play_current_turn method to play a turn.
Then it returns the print_guesser_victory_message when the
guesser_victory? method returns true. Then it calls the set_next_turn
method to prepare the next turn / iteration. Finally, it returns the
print_guesser_defeat_message when the guesser_defeat? method returns
true.