Add calculate_colors_guessed method
This method has one parameter: creator_code. It uses a loop to iterate
through each element of both creator_code and code. For each iteration,
when both colors are the same, it calls the add_color method. And when
the colors are not equal but the color of the code variable is part of
the creator_code, then we also call the add_color method.